requirements.txt
fastapi
uvicorn
python-docx
pytesseract
Pillow
python-multipart
requests
python-dotenv
pymupdf   
streamlit
